remove not useful tests; move github.ts to src/
