--- src/github/__assets__/getPRsLabels.rec
@@ -1,8 +1,0 @@
-https://api.github.com/repos/corentinmusard/otel-cicd-action/issues/18/labels
-https://api.github.com/repos/corentinmusard/otel-cicd-action/issues/18/labels
-200
-W3siaWQiOjY1NzA0MjU5NzUsIm5vZGVfaWQiOiJMQV9rd0RPTFRpOGtzOEFBQUFCaDZDLWR3IiwidXJsIjoiaHR0cHM6Ly9hcGkuZ2l0aHViLmNvbS9yZXBvcy9jb3JlbnRpbm11c2FyZC9vdGVsLWNpY2QtYWN0aW9uL2xhYmVscy9lbmhhbmNlbWVudCIsIm5hbWUiOiJlbmhhbmNlbWVudCIsImNvbG9yIjoiYTJlZWVmIiwiZGVmYXVsdCI6dHJ1ZSwiZGVzY3JpcHRpb24iOiJOZXcgZmVhdHVyZSBvciByZXF1ZXN0In0seyJpZCI6NzA3NDExMDQ1OSwibm9kZV9pZCI6IkxBX2t3RE9MVGk4a3M4QUFBQUJwYVpiLXciLCJ1cmwiOiJodHRwczovL2FwaS5naXRodWIuY29tL3JlcG9zL2NvcmVudGlubXVzYXJkL290ZWwtY2ljZC1hY3Rpb24vbGFiZWxzL3Rlc3QiLCJuYW1lIjoidGVzdCIsImNvbG9yIjoiOTM4MkY5IiwiZGVmYXVsdCI6ZmFsc2UsImRlc2NyaXB0aW9uIjoiIn1d
-https://api.github.com/repos/corentinmusard/otel-cicd-action/issues/19/labels
-https://api.github.com/repos/corentinmusard/otel-cicd-action/issues/19/labels
-200
-W10=
